dedup: extract shared XmlDocText helpers
GetNodeTextWithRefs (cref/paramref flattening), SimplifyTypeName, NormalizeWhitespace
and NormalizePath were ~100 lines copied verbatim between DocCommentParser (Metadata,
parses /// source comments) and XmlDocFileParser (CLI, parses compiler .xml files), so
any cref-rendering fix had to be made twice or `api --docs` output would differ by
source. Move them into one DotnetInspector.Metadata.XmlDocText class that both call;
drop each parser's now-unused WhitespaceRegex/partial/imports.
