+++ src/dotnet-inspect/Inspectors/XmlDocFileParser.cs
@@ -1,5 +1,3 @@
-using System.Text;
-using System.Text.RegularExpressions;
 using System.Xml;
 using DotnetInspector.Metadata;
 
@@ -9,7 +7,7 @@ namespace DotnetInspector.Inspectors;
 /// Parses XML documentation files (.xml) that accompany .NET assemblies.
 /// These files are generated by the compiler and contain structured documentation.
 /// </summary>
-public partial class XmlDocFileParser
+public class XmlDocFileParser
 {
     private readonly Dictionary<string, XmlNode> _members = new(StringComparer.Ordinal);
     // Maps a member's base key (the doc key up to the first '(' or '`', e.g. "M:Ns.Type.Method")
@@ -496,7 +494,7 @@ public partial class XmlDocFileParser
                 var name = param.Attributes?["name"]?.Value;
                 if (name != null)
                 {
-                    parameters[name] = NormalizeWhitespace(GetNodeTextWithRefs(param));
+                    parameters[name] = XmlDocText.NormalizeWhitespace(XmlDocText.GetNodeTextWithRefs(param));
                 }
             }
         }
@@ -516,106 +514,8 @@ public partial class XmlDocFileParser
         if (node == null)
             return null;
 
-        var text = GetNodeTextWithRefs(node);
-        text = NormalizeWhitespace(text);
+        var text = XmlDocText.NormalizeWhitespace(XmlDocText.GetNodeTextWithRefs(node));
         return string.IsNullOrWhiteSpace(text) ? null : text;
-    }
-
-    /// <summary>
-    /// Recursively extracts text from a node, replacing see/paramref/typeparamref
-    /// elements with their referenced names.
-    /// </summary>
-    private static string GetNodeTextWithRefs(XmlNode node)
-    {
-        var sb = new StringBuilder();
-
-        foreach (XmlNode child in node.ChildNodes)
-        {
-            switch (child.NodeType)
-            {
-                case XmlNodeType.Text:
-                case XmlNodeType.Whitespace:
-                case XmlNodeType.SignificantWhitespace:
-                    sb.Append(child.Value);
-                    break;
-
-                case XmlNodeType.Element:
-                    switch (child.Name)
-                    {
-                        case "see":
-                        case "seealso":
-                            var cref = child.Attributes?["cref"]?.Value;
-                            if (cref != null)
-                            {
-                                sb.Append(SimplifyTypeName(cref));
-                            }
-                            else
-                            {
-                                var innerText = child.InnerText;
-                                if (!string.IsNullOrWhiteSpace(innerText))
-                                {
-                                    sb.Append(innerText);
-                                }
-                            }
-                            break;
-
-                        case "paramref":
-                        case "typeparamref":
-                            var name = child.Attributes?["name"]?.Value;
-                            if (name != null)
-                            {
-                                sb.Append(name);
-                            }
-                            break;
-
-                        case "c":
-                            sb.Append(child.InnerText);
-                            break;
-
-                        default:
-                            sb.Append(GetNodeTextWithRefs(child));
-                            break;
-                    }
-                    break;
-            }
-        }
-
-        return sb.ToString();
-    }
-
-    /// <summary>
-    /// Simplifies a cref value to just the type/member name.
-    /// </summary>
-    private static string SimplifyTypeName(string cref)
-    {
-        // Remove type prefix (T:, M:, P:, F:, E:, N:)
-        if (cref.Length > 2 && cref[1] == ':')
-        {
-            cref = cref[2..];
-        }
-
-        // Remove method parameters
-        var parenIndex = cref.IndexOf('(');
-        if (parenIndex > 0)
-        {
-            cref = cref[..parenIndex];
-        }
-
-        // Get just the last part (simple name)
-        var lastDot = cref.LastIndexOf('.');
-        if (lastDot >= 0)
-        {
-            cref = cref[(lastDot + 1)..];
-        }
-
-        // Handle generic arity (e.g., List`1 -> List)
-        var backtick = cref.IndexOf('`');
-        if (backtick >= 0)
-        {
-            cref = cref[..backtick];
-        }
-
-        return cref;
     }
 
     /// <summary>
@@ -639,7 +539,7 @@ public partial class XmlDocFileParser
                 var title = codeNode.Attributes?["title"]?.Value;
 
                 samples.Add(new DocCommentParser.SampleReference(
-                    NormalizePath(source),
+                    XmlDocText.NormalizePath(source),
                     title,
                     region
                 ));
@@ -649,11 +549,4 @@ public partial class XmlDocFileParser
         return samples.Count > 0 ? samples : null;
     }
 
-    private static string NormalizePath(string path) => path.Replace('\\', '/');
-
-    [GeneratedRegex(@"\s+")]
-    private static partial Regex WhitespaceRegex();
-
-    private static string NormalizeWhitespace(string text) =>
-        WhitespaceRegex().Replace(text.Trim(), " ");
 }
